% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakMonitor.R
\name{peakMonitor}
\alias{peakMonitor}
\title{xcms results peak monitor and shiny application}
\usage{
peakMonitor(xcmsOutput = NULL, obsNames = NULL, metab = NULL,
  qcNames = NULL, ppm = 10, rtdev = 30, maxSignalAtt = 20,
  percBelow = 20, obsTypeStr = rep("unknown_obsType", length(obsNames)),
  obsTypeCol = rep("darkgrey", length(obsNames)), launchApp = TRUE,
  analysisDir = NULL)
}
\arguments{
\item{xcmsOutput}{character or data.frame. Full path to xcms results file (or an XCMS output file peak table in the form of a data.frame) if argument is not supplied a tcltk window will open.}

\item{obsNames}{character vector of sample/QC/Blank names to identify these columns in the xcms output table. The obsNames argument should be pre-sorted in to the acquisition run order.}

\item{metab}{character or data.frame. Full path to a csv file of a list of metabolites you wish to monitor (or a data.frame) if argument is not supplied a tcltk window will open.
This list will be used to identify peak groups in the xcms output table within
user defined ppm mass accuracy and retention time deviation parameters.
The table must consist of the following 3 columns minimum

\enumerate{
\item name -- unique name of compound
\item mzmed -- expected mass-to-charge ratio
\item rtmed -- expected retention time (in seconds)
}}

\item{qcNames}{character vector of analytical replicate local QC names to identify these columns in the xcms output table. The qcNames argument should be pre-sorted in to the acquisition run order. These identificatory sample names are critical for CV\% calculation for example.}

\item{ppm}{numeric ppm mass accuracy for matching monitored metabolites
to the xcms output table (default = 10 ppm).}

\item{rtdev}{numeric maximum deviation in seconds for matching monitored
metabolite list and the xcms output table peak groups (default = 30 seconds).}

\item{maxSignalAtt}{numeric percentage (i.e. values between 0-100\%, default=20\%)
maximum percentage last quality control signal attenuation from the quality controls median value.
If a monitored metabolite drops below this value then (i.e. -20\%) then it will be
flagged in the returned table.}

\item{obsTypeStr}{character vector equal in length to obsNames to identify different run types
in the obsNames argument.}

\item{obsTypeCol}{character vector colours equal in length to obsNames to display for run types in application.}

\item{launchApp}{logical default is true. If true peakMonitor shiny app will open in your default web browser.}

\item{analysisDir}{the full path of the peakMonitor directory (output/peakMonitor)
containing the output files (.csv) from \code{\link{peakMonitor}}. This
argument is only required if the argument launchApp is TRUE. If the argument
is not supplied but required then a tcltk window will open and the directory
can be selected.}

\item{maxBelow}{numeric (values between 0-100\%, default = 20 \%) maximum percentage of monitored
metabolites below the maximum signal attenutation argument (see argument maxSignalAtt). E.g. if 20\% of monitored metabolites contained in expected metabolite table have dropped
below 20\% median QC signal signal attenuation.}
}
\value{
a list containing two elements named: \enumerate{
\item "monitMetab" a data.frame containing the metabolite peak monitoring results.
\item "percBelow" a logical whether the user parameterized maximum percentage of
monitored metabolites were below the maximum signal attenuation value. This value
is used by simExTargId to notify the user of any significant signal attenuation drop during
a run for example.
}
}
\description{
xcms results peak monitor and shiny application
}
