% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawFileConvert.R
\name{rawFileConvert}
\alias{rawFileConvert}
\title{Convert proprietory instrument manufacturer file types to .mzXML files}
\usage{
rawFileConvert(rawFiles = NULL, mzXmlDir = NULL, nSlaves = NULL,
  subSetSecs = NULL, zlib = T, centroid = TRUE, MS2 = FALSE)
}
\arguments{
\item{rawFiles}{character vector of full path names to raw data files.}

\item{mzXmlDir}{character full path to directory into which to write the mzXml
files output by MSConvert.}

\item{nSlaves}{numeric number of computer cores for parallel computation.}

\item{subSetSecs}{numeric vector of a minimum and maximum time window in 
mzXML files (e.g. c(1500, 1900)).}

\item{zlib}{logical whether or not to apply zlib compression (default = TRUE). Will reduce
file size but may affect downstream data analysis.}

\item{centroid}{do the raw data files need to be centroided during conversion by MSConvert
\url{http://proteowizard.sourceforge.net/downloads.shtml}. NB. centroiding of data
is necessary for utilization of the "centWave" algorithm of xcms (\code{\link{findPeaks.centWave-methods}}).}

\item{MS2}{logical should raw LC-MS data files converted by MSConvert to mzXml be converted
as MS/MS files (e.g. data-dependent MS/MS files).}
}
\description{
Convert proprietory instrument manufacturer file types to .mzXML files
}
\details{
MSConvert (ProteoWizard \url{http://proteowizard.sourceforge.net/downloads.shtml}) 
must be installed and in the case of windows
in the path in the environmental variables. The mzXML file conversion will occur
through shell commands using the command line version of MSConvert. If more than
one cpu is utilized the mzXML file conversion process will occur as a parallel
computation.
}

