% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCsvAttempt.R
\name{writeCsvAttempt}
\alias{writeCsvAttempt}
\title{try to write a csv and pause and alert user if open}
\usage{
writeCsvAttempt(df = NULL, fileName = NULL)
}
\arguments{
\item{df}{dataframe}

\item{fileName}{complete path to new .csv file}
}
\description{
try to write a csv and pause and alert user if open
}
