% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simExTargId.R
\name{simExTargId}
\alias{simExTargId}
\title{simExTargId}
\usage{
simExTargId(rawDir = NULL, analysisDir = NULL, coVar = NULL,
  nSlaves = NULL, minFiles = 10, centroid = TRUE, MS2 = FALSE,
  zeroFillvalue = NULL, normMethod = NULL, manBatchAdj = NULL,
  LogTransBase = exp(1), blankFC = 2, replicates = FALSE,
  pAdjustMethod = "BH", corrThresh = 0.8, minFeat = 1,
  hclustMethod = "median", distMeas = "euclidean", bw = 2,
  mzwid = 0.015, minfrac = 0.25, xcmsSetArgs = list(peakwidth = c(2, 20),
  ppm = 10, snthresh = 5, method = "centWave"), pcaOutIdArgs = list(cv = "q2",
  scale = "pareto", centre = T), maxTime = 60)
}
\arguments{
\item{rawDir}{character full path name of raw data directory into which raw data will 
eventually/ has already been started to be written. An "_analysis" directory
will be automatically created into which all converted files (.mzXML) and
results output will be saved. Fifteen minutes after the last modification of a new
raw data file a conversion to the mzXML format and initial peak-picking will
occur (\code{\link{xcmsSet}}).}

\item{analysisDir}{full path name of analysis directory, all mzXML, peak-picking
and results files will be saved here.}

\item{coVar}{covariate table can be supplied as either a full path name to a 
.csv file (see details) or as a \code{\link{data.frame}}.}

\item{nSlaves}{numeric number of computer cores for parallel computation.}

\item{minFiles}{the minimum number of raw data file converted to mzXML files
before commencing subsequent steps of XCMS processing, pre-processing, PCA and
statistical analysis. Default = 10.}

\item{centroid}{do the raw data files need to be centroided during conversion by MSConvert
\url{http://proteowizard.sourceforge.net/downloads.shtml}. NB. centroiding of data
is necessary for utilization of the "centWave" algorithm of xcms (\code{\link{findPeaks.centWave-methods}}).}

\item{MS2}{logical should raw LC-MS data files converted by MSConvert to mzXml be converted
as MS/MS files (e.g. data-dependent MS/MS files).}

\item{zeroFillvalue}{value to fill missing and zero values in the xcms peak table. If argument is left blank
the default is to fill with half the smallest non-zero value. (see: \code{\link{zeroFill}}).}

\item{normMethod}{normalization method to perform (see: \code{\link{signNorm}}). 
If argument is left blank no normalization is performed. Options include 
median fold change (probabilistic quotient) 'medFC' or total ion signal 
'totIon' methods.}

\item{manBatchAdj}{character vector of co-variate table column names. If argument is not supplied automatic cluster identification/ batch adjustment using the \code{\link{pcaClustId}} function will occur prior to PCA analysis. If multiple
column names are supplied a multiple linear regression will be calculated and the batch adjusted residuals obtained from the model.}

\item{LogTransBase}{numeric base value for log transformation. defaults to exponential of 1 (see: \code{\link{logTrans}}).}

\item{blankFC}{numeric minimum fold difference for each xcms peak table feature 
between the sample injections (numerator) and negative control blank injections (denominator). 
This will result in all xcms peak table features which are less than blankFC 
fold change higher in the samples than the blanks being removed (default = 2 fold).
This background substraction will not occur until there are at least 1 blank and 1 sample
data file acquired and converted to mzXML files.}

\item{replicates}{logical (default = FALSE) if TRUE then the 3rd column of the co-variate table supplied will be used to identify analytical/ preparative replicates of the same sample. This information will be used to average signal intensities of analytical replicates.}

\item{pAdjustMethod}{character p-value multiple testing adjustment method (see: \code{\link{p.adjust}}).}

\item{corrThresh}{correlation coefficient threshold (non-parametric Spearman's 
Rho) to group features within a retention time cluster.}

\item{minFeat}{minimum number of features with a Retention time/ correlation
cluster to consider it a group (default = 2, i.e. a cluster must contain at
least 2 features to be considered a group).}

\item{hclustMethod}{hierarchical clustering method to \code{\link{hclust.vector}} 
method of fastcluster package (default = "median").}

\item{distMeas}{distance measure for retention time clustering (default = "euclidean"). 
see \code{\link{hclust.vector}}.
Within retention clusters dissimilarity is computed as 1-correlation coefficient.}

\item{xcmsSetArgs}{list of arguments of the xcms function \code{\link{xcmsSet}}.}

\item{pcaOutIdArgs}{list of arguments to the \code{\link{pcaOutId}} function.}

\item{maxTime}{maximum time (in minutes) from the time the last raw data file
was written. If most recent raw data file is older than this time then simExTargId
will stop. This is designed to stop the process if necessary after an extended period
of time. default = 60 mins.}
}
\description{
This function seeks to remove the gap between metabolomic 
MS1 profiling experiments and discovery of statistically relevant targets for 
MSn fragmentation based identification.
}
\details{
The function is designed to faciliate 
simultaneous collection of raw metabolomic
profiling data, conversion from an instrument manufacturer's proprietary format to the
 open file format mzXML conversion (using the command line
version of MSConvert) to a new analysis directory, xcms based peak picking/
alignment, data pre-processing (\code{\link{zeroFill}}, \code{\link{logTrans}},
 \code{\link{signNorm}}), automatic PCA-based outlier removal, scores plot
 cluster identification and automatic potential batch adjustment (\code{\link{pcaOutId}}, 
 \code{\link{pcaClustId}}, \code{\link{batchAdj}}), automatic co-variate based
 statistical analysis and feature deconvolution (\code{\link{coVarStatType}},
  \code{\link{rtCorrClust}}). The data output of the workflow can be 
  visualized by a shiny application (\code{\link{shiny}}) at any time 
  (\code{\link{targIdShiny}}) and MSn fragmentation targets rapidly identified. 

 The function works according to the following process:

 1. The function can be run before the collection of the first raw data file
 or after collection of any number of raw MS data files. 
 
 2. The directory location where the raw data are being written to must be provided
 (rawDir), as well as a comma delimited text file (.csv) containing two named columns.
 The first column must contain the precise names of each raw MS data file
 that will be eventually created in the raw data directory. 
 The second column of the .csv file must contain the binary class membership of
 each raw MS data file to be considered for statistical analysis. e.g. case/ control,
 treated/ untreated, exposed/ unexposed, blank/ sample 
 
 N.B. Raw data files such as blank and quality control for example which are written to 
 the raw data directory during an experiment will also be included in the
 xcms peak-picking and alignment process but will not be considered in the 
 subsequent statistical analysis.
 
 This information will be used to inform the 
 non-parametric statistical analysis at present only a binary case is supported
 i.e. two sample classes.    


analysis.
}
\examples{
# example XCMS peak picking/ grouping/ alignment settings for high-resolution LC-MS data (Q-ToF). 
# peakwidth=c(2, 20), ppm=10, snthresh=5, bw=2, mzwid=0.015, minfrac=0.25, method="centWave"
}
\seealso{
\code{\link{xcmsSet}}, \code{\link{retcor}}, \code{\link{group}}, 
\code{\link{fillPeaks}}, \code{\link{diffreport}},
}

